// src/api/openapi/models/UpdateAccessControl.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateAccessControl = {
    farms?: boolean;
    learning?: boolean;
    event?: boolean;
    blog?: boolean;
    forums?: boolean;
    admin?: boolean;
    cuai?: boolean;
    home?: boolean;
    about?: boolean;
    users?: boolean;
    privacy_policy?: boolean;
    terms_and_conditions?: boolean;
    user_feedback?: boolean;
    crops?: boolean;
    help_center?: boolean;
    activity_logs?: boolean;
};